refactor: 테스트 픽스처 패키지 이동
