+++ README.md
@@ -86,4 +86,4 @@
 
 ### 3단계 피드백
 - [x] ControllerScanner 가 불필요한 인스턴스 필드를 가지지 안도록 변경(유틸성으로 변경하여 정적 메서드로 컨스톨러 인스턴스들을 반환)
-- [ ] 테스트 컨트롤러 픽스처의 의도를 표현할 수 있도록 픽스처 패키지로 이동 
+- [x] 테스트 컨트롤러 픽스처의 의도를 표현할 수 있도록 픽스처 패키지로 이동
