+++ src/test/java/next/fixture/controller/NamedAnnotatedController.java
@@ -1,4 +1,4 @@
-package next.controller;
+package next.fixture.controller;
 
 import core.annotation.web.Controller;
