+++ src/test/java/next/fixture/duplicatedcontroller/HomeController.java
@@ -1,4 +1,4 @@
-package next.duplicatedcontroller;
+package next.fixture.duplicatedcontroller;
 
 import core.annotation.web.Controller;
